Implemented Disease Prediction Using Roboflow

+++ backend/app/schemas/disease.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class DiseaseResponse(BaseModel):
     disease: str
-    confidence: int
+    confidence: float
     low_confidence: bool
     recommended_actions: list[str]
     language: str
